Coupons Rest client and other refactorings
