+++ src/main/java/cl/talavera/api/ApiApplication.java
@@ -2,6 +2,9 @@ package cl.talavera.api;
 
 import cl.talavera.api.core.domain.Category;
 import cl.talavera.api.data.CategoryResponse;
+import cl.talavera.api.data.CategoryRestRepository;
+import cl.talavera.api.data.CouponResponse;
+import cl.talavera.api.data.CouponRestRepository;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -10,6 +13,7 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.context.annotation.Bean;
+import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestTemplate;
 
 import java.util.List;
@@ -23,6 +27,15 @@ public class ApiApplication {
 		SpringApplication.run(ApiApplication.class, args);
 	}
 
+}
+
+@Component
+class ApplicationRunner implements CommandLineRunner{
+	@Override
+	public void run(String... args) throws Exception {
+		//use this to run functional test!!
+		System.out.println("Welcome!!!");
+	}
 }
