+++ src/main/java/cl/talavera/api/core/port/category/CategoriesPresenterPort.java
@@ -1,7 +1,7 @@
 package cl.talavera.api.core.port.category;
 
 import cl.talavera.api.core.domain.Category;
-import cl.talavera.api.rest.CategoryView;
+import cl.talavera.api.web.CategoryView;
 
 import java.util.List;
