+++ src/main/java/cl/talavera/api/data/CouponResponse.java
@@ -1,0 +1,11 @@
+package cl.talavera.api.data;
+
+import lombok.Data;
+
+import java.util.Date;
+
+@Data
+public class CouponResponse {
+    private String id;
+    private Date expiresAt;
+}
