+++ src/main/java/cl/talavera/api/web/TopCategoriesEndpoint.java
@@ -1,4 +1,4 @@
-package cl.talavera.api.rest;
+package cl.talavera.api.web;
 
 import cl.talavera.api.core.port.category.CategoriesPresenterPort;
 import cl.talavera.api.core.port.category.TopCategoriesInteractorPort;
